+++ src/realm/parser/property_expression.cpp
@@ -41,7 +41,12 @@ PropertyExpression::PropertyExpression(Query& q, const std::string& key_path_str
                                             element.table->get_column_name(element.col_key),
                                             get_printable_table_name(*element.table)));
             if (element.table == cur_table) {
-                cur_table = element.table->get_link_target(element.col_key); // advance through forward link
+                if (element.col_key == ColKey()) {
+                    cur_table = element.table;
+                }
+                else {
+                    cur_table = element.table->get_link_target(element.col_key); // advance through forward link
+                }
             }
             else {
                 cur_table = element.table; // advance through backlink